Implementation of InsMove
Added an iteractables dictionary to the level so we can access them.

Fixed the file path to the tile set.

Added the name and access_side to the Desk.

Added handling to start calling the user's Instructions.

Implemented the InsMove function to teleport the actor to the destination (Desk only right now).